[Add] #191 - accesstoken userId 일치 여부 검증 및 정보 수정시 refreshtoken 검증 추가

+++ src/main/java/shop/cazait/domain/user/api/UserApiController.java
@@ -1,5 +1,6 @@
 package shop.cazait.domain.user.api;
 
+import static shop.cazait.global.error.status.ErrorStatus.INVALID_REQUEST;
 import static shop.cazait.global.error.status.SuccessStatus.CREATE_USER;
 import static shop.cazait.global.error.status.SuccessStatus.SUCCESS;
 
@@ -10,6 +11,7 @@ import io.swagger.annotations.ApiOperation;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
+import java.util.List;
 import javax.crypto.BadPaddingException;
 import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
@@ -18,25 +20,14 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import shop.cazait.domain.user.dto.DeleteUserRes;
-import shop.cazait.domain.user.dto.PatchUserReq;
-import shop.cazait.domain.user.dto.PatchUserRes;
-import shop.cazait.domain.user.dto.PostUserReq;
-import shop.cazait.domain.user.dto.PostUserRes;
+import org.springframework.web.bind.annotation.*;
+import shop.cazait.domain.user.dto.*;
 import shop.cazait.domain.user.exception.UserException;
 import shop.cazait.domain.user.service.UserService;
 import shop.cazait.global.common.dto.response.SuccessResponse;
 import shop.cazait.global.config.encrypt.JwtService;
 import shop.cazait.global.config.encrypt.NoAuth;
+import shop.cazait.global.error.status.ErrorStatus;
 
 @Api(tags = "유저 API")
 @Validated
@@ -57,21 +48,27 @@ public class UserApiController {
         return new SuccessResponse<>(CREATE_USER, postUserRes);
     }
 
-//    @GetMapping("/all")
-//    @ApiOperation(value = "모든 회원을 조회",notes = "회원가입된 모든 회원 정보를 조회")
-//    public SuccessResponse<List<GetUserRes>> getUsers(){
-//        List<GetUserRes> allGetUserRes = userService.getAllUsers();
-//        return new SuccessResponse<>(allGetUserRes);
-//    }
+    @GetMapping("/all")
+    @ApiOperation(value = "모든 회원을 조회",notes = "회원가입된 모든 회원 정보를 조회")
+    public SuccessResponse<List<GetUserRes>> getUsers(){
+        List<GetUserRes> allGetUserRes = userService.getAllUsers();
+        return new SuccessResponse<>(SUCCESS, allGetUserRes);
+    }
 
-//    @GetMapping("/{userIdx}")
-//    @ApiOperation(value = "특정 회원 정보를 조회", notes ="자신의 계정 정보를 조회")
-//    @ApiImplicitParam (name="userIdx",value = "사용자 userId")
-//    public SuccessResponse<GetUserRes> getUser(
-//             @PathVariable(name = "userIdx") Long userIdx) throws UserException {
-//        GetUserRes userInfoRes = userService.getUserInfo(userIdx);
-//        return new SuccessResponse<>(userInfoRes);
-//    }
+    @GetMapping("/{userIdx}")
+    @ApiOperation(value = "특정 회원 정보를 조회", notes ="자신의 계정 정보를 조회")
+    @ApiImplicitParam (name="userIdx",value = "사용자 userId")
+    public SuccessResponse<GetUserRes> getUser(
+             @PathVariable(name = "userIdx") Long userIdx) throws UserException {
+        Long userIdxFromJwt = jwtService.getUserIdx();
+
+        if(!userIdx.equals(userIdxFromJwt)){
+            throw new UserException(INVALID_REQUEST);
+        }
+
+        GetUserRes userInfoRes = userService.getUserInfo(userIdx);
+        return new SuccessResponse<>(SUCCESS, userInfoRes);
+    }
 
     @PatchMapping("/{userIdx}")
     @ApiOperation(value="특정한 회원 정보를 수정", notes = "자신의 계정 정보를 수정")
@@ -82,15 +79,29 @@ public class UserApiController {
     public SuccessResponse<PatchUserRes> modifyUser(
             @PathVariable(name = "userIdx") Long userIdx,
             @RequestBody @Valid  PatchUserReq patchUserReq,
-            @RequestHeader(value="REFRESH-TOKEN") String refreshToken) {
-        PatchUserRes patchUserRes = userService.modifyUser(userIdx, patchUserReq, refreshToken);
-        return new SuccessResponse<>(SUCCESS, patchUserRes);
+            @RequestHeader(value="REFRESH-TOKEN") String refreshToken) throws UserException {
+        if(jwtService.isValidRefreshToken(refreshToken)){
+            Long userIdxFromJwt = jwtService.getUserIdx();
+
+            if (!userIdx.equals(userIdxFromJwt)) {
+                throw new UserException(INVALID_REQUEST);
+            }
+        }
+
+            PatchUserRes patchUserRes = userService.modifyUser(userIdx, patchUserReq, refreshToken);
+            return new SuccessResponse<>(SUCCESS, patchUserRes);
+
     }
 
     @DeleteMapping("/{userIdx}")
     @ApiOperation(value="특정한 회원 정보를 삭제", notes = "자신의 계정 정보를 삭제")
     @ApiImplicitParam (name="userIdx",value = "사용자 userId")
-    public SuccessResponse<DeleteUserRes> deleteUser(@PathVariable(name = "userIdx") Long userIdx) {
+    public SuccessResponse<DeleteUserRes> deleteUser(@PathVariable(name = "userIdx") Long userIdx) throws UserException {
+        Long userIdxFromJwt = jwtService.getUserIdx();
+
+        if(!userIdx.equals(userIdxFromJwt)){
+            throw new UserException(INVALID_REQUEST);
+        }
 
         DeleteUserRes deleteUserRes = userService.deleteUser(userIdx);
         return new SuccessResponse<>(SUCCESS, deleteUserRes);
